# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API сервис, который является частью системы геооповещений",
        "title": "geonotify-service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/incidents": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить все инциденты с поддержкой пагинации",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Получить список инцидентов с пагинацией (оператор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы (по умолчанию 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит на страницу (по умолчанию 10, максимум 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_resp.IncidentsListResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры пагинации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать новую опасную зону (требуется API key)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Создать инцидент (оператор)",
                "parameters": [
                    {
                        "description": "Данные инцидента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_req.IncidentCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_resp.IncidentCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/incidents/{incident_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Детали конкретной зоны опасности",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Получить инцидент по ID (оператор)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID инцидента",
                        "name": "incident_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_resp.IncidentResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Инцидент не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Полное обновление данных существующей опасной зоны (PUT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Обновить инцидент (оператор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID инцидента",
                        "name": "incident_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Полные данные инцидента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_req.IncidentUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Инцидент обновлен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Инцидент не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Мягкое удаление опасной зоны",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Удалить инцидент (оператор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID инцидента",
                        "name": "incident_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Инцидент удален",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Инцидент не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_4otis_geonotify-service_internal_dto_req.IncidentCreateRequest": {
            "type": "object",
            "properties": {
                "descr": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "radius_m": {
                    "type": "number"
                }
            }
        },
        "github_com_4otis_geonotify-service_internal_dto_req.IncidentUpdateRequest": {
            "type": "object",
            "properties": {
                "descr": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "radius_m": {
                    "type": "number"
                }
            }
        },
...
        "github_com_4otis_geonotify-service_internal_dto_resp.IncidentCreateResponse": {
            "type": "object",
            "properties": {
                "incident_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_4otis_geonotify-service_internal_dto_resp.IncidentResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "descr": {
                    "type": "string"
                },
                "incident_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "radius_m": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_4otis_geonotify-service_internal_dto_resp.IncidentsListResponse": {
            "type": "object",
            "properties": {
                "incidents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_4otis_geonotify-service_internal_dto_resp.IncidentResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}